+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 * STDIN pipe/stream will upload the chunks in oneshot not using `Transfer-Encoding: chunked`
+* `{bytes}/{total_bytes}` in progressbar.
 
 
 ## 0.3.0  🚲